... with the right irq mask
It's the NVIC mask, not the exception mask

+++ core/MemoryBus.h
@@ -117,7 +117,7 @@ public:
     }
 
     void peripheralUpdate(uint64_t target);
-    void peripheralUpdate(uint64_t target, uint64_t irqMask);
+    void peripheralUpdate(uint64_t target, uint32_t irqMask);
     uint64_t getNextInterruptTime(uint64_t target);
 
     void setPendingIRQ(int n);
